eKYC/CMND-CCCD/requirements.txt
pillow
tensorflow
matplotlib
PyYAML
opencv-python
numpy
vietocr
torch
torchvision